scripts/requirements.txt
# Works with Python 3.6+
spotipy>=2.23.0
python-dotenv>=1.0.0
yt-dlp>=2023.0.0
